backend/requirements.txt
psycopg2-binary
pydantic~=2.9.2
